Decouple shared scenarioeditor files from OpenLayers
- Split featureLayerUtils.ts: OL selection/sync/feature code moves to
  featureLayerUtilsOl.ts; the neutral icon/type helpers stay put
- referenceFeatureUtils.ts uses local structural types instead of OL
  FeatureLike/BaseLayer
- MapEditorMeasurementToolbar.vue gets the native map via the MapAdapter
  (getNativeMap) instead of the OL-typed inject
- OL-typed injection keys move from components/injects.ts to
  scenarioeditor/olInjects.ts
- Shrink PENDING_OPENLAYERS_DECOUPLE accordingly
